Update folder structure
Since we use this as a starting point for real projects it is really stange to have
e.g. some data in `/data` and some in `shared/data` so just get rif of the shared folder.

For components I think it is usefull to keep it around because shared components
all need to have a stories.js file.
Also hopefully shrared comonents will only become more and more and that might
makes it hard to find a component of the current project